Improved MusicRoomFragment: Separated Into Owner and Listener
